refactor(server/auth): removed sign-in service file and wrote in
